+++ CHANGELOG.md
@@ -9,7 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added support for YAML, MsgPack, and Pickle formats, behind the `extra_formats` feature flag
 
-## v0.4.0 - 2020-08-18
+## v0.4.0-rc2 - 2020-08-27
+
+### Added
+
+- Added a `reconnect` method on the `Broker`.
+
+### Changed
+
+- `Celery` and `Beat` apps will automatically try to reconnect the broker when the connection fails.
+
+## v0.4.0-rc1 - 2020-08-18
 
 ### Added